[lib-react] Unbreak material-ui theming
Temporarily by a hack. Something might be wrong with Yarn, next, or material-ui.
