+++ packages/lib-react/MaterialThemedApp.tsx
@@ -1,15 +1,32 @@
-import React, { ReactElement, ComponentProps } from 'react';
+import React, { ReactElement, ReactNode } from 'react';
 
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
+import teal from '@material-ui/core/colors/teal';
+import { MuiThemeProvider, StylesProvider, createMuiTheme } from '@material-ui/core/styles';
 
-import MaterialPreConfiguredThemedApp from './MaterialPreConfiguredThemedApp';
+const theme = createMuiTheme({
+  palette: {
+    primary: { main: '#3E7AE2' },
+    secondary: { main: teal[500] },
+  },
+});
 
-type MaterialPreConfiguredThemedAppPassedThroughProps = Omit<
-  ComponentProps<typeof MaterialPreConfiguredThemedApp>,
-  'toolBarChildren'
->;
+type AppBarPosition = 'fixed' | 'absolute' | 'sticky' | 'static' | 'relative';
 
-type Props = MaterialPreConfiguredThemedAppPassedThroughProps & {
+type StyleProps = {
+  readonly app?: string;
+  readonly appBar?: string;
+  readonly title?: string;
+};
+
+type Props = {
+  readonly appBarPosition?: AppBarPosition;
+  readonly styles: StyleProps;
+  readonly toolBarChildren: ReactElement;
+  readonly buttons?: ReactElement;
+  readonly children: ReactNode;
   readonly title: string;
 };
 
@@ -20,23 +37,23 @@ const App = ({
   buttons,
   children,
 }: Props): ReactElement => {
-  const toolBarChildren = (
-    <Typography variant="h6" color="inherit" className={styles.title}>
-      {title}
-    </Typography>
-  );
   return (
-    <MaterialPreConfiguredThemedApp
-      appBarPosition={appBarPosition}
-      styles={styles}
-      toolBarChildren={toolBarChildren}
-      buttons={buttons}
-    >
-      {children}
-    </MaterialPreConfiguredThemedApp>
+    <MuiThemeProvider theme={theme}>
+      <StylesProvider injectFirst>
+        <div className={styles.app}>
+          <AppBar position={appBarPosition} className={styles.appBar}>
+            <Toolbar>
+              <Typography variant="h6" color="inherit" className={styles.title}>
+                {title}
+              </Typography>
+              {buttons}
+            </Toolbar>
+          </AppBar>
+          {children}
+        </div>
+      </StylesProvider>
+    </MuiThemeProvider>
   );
 };
-
-App.defaultProps = { appBarPosition: 'static' };
 
 export default App;
