+++ packages/lib-react/package.json
@@ -22,7 +22,6 @@
     "react-dom": "^16.13.1"
   },
   "devDependencies": {
-    "@material-ui/core": "^4.9.14",
     "@types/jest": "^25.2.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
