+++ packages/samlang/package.json
@@ -15,7 +15,7 @@
     "@dev-sam/samlang-demo": "^0.0.2",
     "@docusaurus/core": "^2.0.0-alpha.55",
     "@docusaurus/preset-classic": "^2.0.0-alpha.55",
-    "@material-ui/core": "^4.9.14",
+    "@material-ui/core": "^4.10.0",
     "autosize": "^4.0.2",
     "classnames": "^2.2.6",
     "lib-docusaurus-plugin": "0.1.0",
