+++ packages/ten/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.9.14",
+    "@material-ui/core": "^4.10.0",
     "lib-react": "0.1.0",
     "next": "^9.4.2",
     "react": "^16.13.1",
