--- packages/www/pages/_app.ts
@@ -1,4 +1,0 @@
-import App from 'next/app';
-import './index.css';
-
-export default App;
