+++ packages/www/pages/_document.tsx
@@ -1,5 +1,8 @@
+// Adapted from: https://github.com/mui-org/material-ui/tree/master/examples/nextjs-with-typescript
+
 import React, { ReactElement } from 'react';
 
+import { ServerStyleSheets } from '@material-ui/core/styles';
 import Document, {
   DocumentContext,
   DocumentInitialProps,
@@ -11,14 +14,33 @@ import Document, {
 
 class MyDocument extends Document {
   static async getInitialProps(ctx: DocumentContext): Promise<DocumentInitialProps> {
+    const sheets = new ServerStyleSheets();
+    const originalRenderPage = ctx.renderPage;
+
+    ctx.renderPage = () =>
+      originalRenderPage({
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
+      });
+
     const initialProps = await Document.getInitialProps(ctx);
-    return { ...initialProps };
+    const styles = [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()];
+    return { ...initialProps, styles };
   }
 
   render(): ReactElement {
     return (
       <Html lang="en">
-        <Head />
+        <Head>
+          <link
+            rel="stylesheet"
+            href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
+          />
+          <link
+            rel="stylesheet"
+            href="https://fonts.googleapis.com/css?family=Roboto+Mono:400,500&display=swap"
+          />
+        </Head>
         <body>
           <Main />
           <NextScript />
